ref(integrations): Separate frontend concerns
Splits the OrganizationIntegrationConfig component into a multiple
components, separating out the quickly-growing component into more
manageable components.

 - Moves the components into the `organizationIntegrations` module.

 - Adds the `IntegrationDescription` component to represent the static
   description of an integration.

 - Adds the `IntegrationInstallation` component to represent the
   integration panel row, and subsequent configuration that will pop
   down from the row.
