+++ src/sentry/static/sentry/app/routes.jsx
@@ -403,7 +403,7 @@ const projectSettingsRoutes = (
       path="integrations/:providerKey/"
       name="Integration Configuration"
       componentPromise={() =>
-        import(/* webpackChunkName: "OrganizationIntegrationConfig" */ './views/organizationIntegrationConfig')}
+        import(/* webpackChunkName: "OrganizationIntegrations" */ './views/organizationIntegrations')}
       component={errorHandler(LazyLoad)}
     />
     <Route
